Introduces a hook to abstract getting tables/columns, updates TypeaheadInput to render a regular input field if no options are passed in.
